+++ build-tools-integration-tests/src/test/java/org/projectnessie/buildtools/TestPublishedPoms.java
@@ -111,8 +111,6 @@ public class TestPublishedPoms {
 
   @ParameterizedTest
   @ValueSource(strings = {
-    "nessie-spark-antlr-runtime",
-    "nessie-spark-extensions-grammar",
     "nessie-spark-extensions-3.3_2.12",
     "nessie-spark-extensions-3.3_2.13",
     "nessie-spark-extensions-3.4_2.12",