+++ integrations/spark-extensions-base/src/main/scala/org/apache/spark/sql/catalyst/plans/logical/CreateReferenceCommand.scala
@@ -19,6 +19,7 @@ import org.apache.spark.sql.catalyst.expressions.Attribute
 
 case class CreateReferenceCommand(
     reference: String,
+    refTimestampOrHash: Option[String],
     isBranch: Boolean,
     catalog: Option[String],
     fromReference: Option[String],